:bug: fix: Send the correct case to ZTC service
For some requests (like POSTS) BFF sends data as-is to the service
without serialising using the proper Struct.

As-is, was actually not as-is, as DRF was translating camelCase to
snake_case and consequently BFF was sending snake_case fields to ZTC
service. OZ uses snake_case internally, so that's why it did still
respond as expected, but that's a coincidence and implementation detail
of OZ we shouldn't depend on.